+++ anonymous_sdk/test_utils_test.go
@@ -1,4 +1,4 @@
-package ssks_tmr
+package anonymous_sdk
 
 import (
 	"github.com/rs/zerolog"