+++ sdk/contacts_test.go
@@ -2,12 +2,12 @@ package sdk
 
 import (
 	"fmt"
+	"github.com/seald/go-seald-sdk/sdk/sigchain"
+	"github.com/seald/go-seald-sdk/test_utils"
+	"github.com/seald/go-seald-sdk/utils"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/ztrue/tracerr"
-	"go-seald-sdk/sdk/sigchain"
-	"go-seald-sdk/test_utils"
-	"go-seald-sdk/utils"
 	"strconv"
 	"testing"
 	"time"