+++ symmetric_key/symmetric_key_test.go
@@ -1,10 +1,10 @@
 package symmetric_key
 
 import (
+	"github.com/seald/go-seald-sdk/test_utils"
+	"github.com/seald/go-seald-sdk/utils"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"go-seald-sdk/test_utils"
-	"go-seald-sdk/utils"
 	"os"
 	"path/filepath"
 	"testing"